+++ CHANGELOG.md
@@ -2,6 +2,26 @@
 
 All Notable changes to `bakame/http-strucured-fields` will be documented in this file.
 
+## [Next] - TBD
+
+### Added
+
+- `InnerList::fromPair`
+- `InnerList::fromAssociative`
+
+### Fixed
+
+- None
+
+### Deprecated
+
+- None
+
+### Removed
+
+- `InnerList::fromPairParameters` use `InnerList::fromPair` instead.
+- `InnerList::fromAssociativeParameters` use `InnerList::fromAssociative` instead.
+
 ## [0.8.0] - 2023-03-12
 
 ### Added